Fix Windows platform plugin narrowing conversion compile errors
Implicit conversions from int to BYTE (unsigned char) result in
compile errors when compiling with the GCC -std=c++0x option.

+++ src/plugins/platforms/windows/qwindowsbackingstore.cpp
@@ -99,7 +99,7 @@ void QWindowsBackingStore::flush(QWindow *window, const QRegion &region,
         SIZE size = {r.width(), r.height()};
         POINT ptDst = {r.x(), r.y()};
         POINT ptSrc = {0, 0};
-        BLENDFUNCTION blend = {AC_SRC_OVER, 0, (int)(255.0 * rw->opacity()), AC_SRC_ALPHA};
+        BLENDFUNCTION blend = {AC_SRC_OVER, 0, (BYTE)(255.0 * rw->opacity()), AC_SRC_ALPHA};
         RECT dirty = {dirtyRect.x(), dirtyRect.y(),
             dirtyRect.x() + dirtyRect.width(), dirtyRect.y() + dirtyRect.height()};
         UPDATELAYEREDWINDOWINFO info = {sizeof(info), NULL, &ptDst, &size, m_image->hdc(), &ptSrc, 0, &blend, ULW_ALPHA, &dirty};
